The User Can Now Choose Meal Times.
The meal time buttons now open up the system time picker dialog to allow the user to set times.

Switching between manual and automatic time selection changes the labels to display the user's manual times or the current automatic time boundaries depending on the setting.

The chosen manual times are not saved currently. There will be checks for overlapping time choices in the future before this.