Added Checks for Overlapping Time Intervals.
When a user sets a time interval for a meal that overlaps with another meal the app will put a warning icon next to the meals that are overlapping.

When the user attempts to go back with overlapping times the app will display a warning that the changes will not be saved.