Local caching of density entries
After density entries are successfully downloaded from the server, they
are cached in the app's cache folder (the same parent directory are the
image files). So, unless it's a fresh install or the cache is cleared,
the getAllDensities api call will not get made.

Also, I added lbs to the list of masses, not sure how I missed that.