Added MealDetailsActivity Tutorial.
Created several tutorial pages to be shown in MealDetailsActivity.

Small changes to MealDetailsActivity layout to remove the padding around the entire activity (it added padding to the dimming effect of the tutorial.)

Added elevation attribute to the tutorial layout so that buttons would not be drawn over it on post-Lollipop devices.